add train file

+++ common/dataset.py
@@ -142,7 +142,7 @@ class AudioDataset(Dataset):
 
         return (segment,
                 torch.tensor(segment.shape[0]).int(),
-                torch.tensor(s["transcript"]),
+                torch.tensor(s["transcript"]).int(),
                 torch.tensor(len(s["transcript"])).int(),
                 s["transcript"])
 
@@ -208,7 +208,7 @@ def collate_fn(batch):
     max_len = lambda l, idx: max(el[idx].size(0) for el in l)
     audio = torch.zeros(bs, max_len(batch, 0))
     audio_lens = torch.zeros(bs, dtype=torch.int32)
-    transcript = torch.zeros(bs, max_len(batch, 2))
+    transcript = torch.zeros(bs, max_len(batch, 2), dtype=torch.int64)
     transcript_lens = torch.zeros(bs, dtype=torch.int32)
 
 
@@ -217,6 +217,7 @@ def collate_fn(batch):
         audio_lens[i] = sample[1]
         transcript[i].narrow(0, 0, sample[2].size(0)).copy_(sample[2])
         transcript_lens[i] = sample[3]
+        
     return audio, audio_lens, transcript, transcript_lens
